+++ test/bigtest/interactors/loan-actions-history.js
@@ -11,6 +11,8 @@ import KeyValue from './KeyValue';
 
   requests = scoped('[data-test-loan-actions-history-requests] div', KeyValue);
   feeFines = scoped('[data-test-loan-fees-fines] [data-test-kv-value]', KeyValue);
+  overduePolicy = scoped('[data-test-overdue-policy] [data-test-kv-value]', KeyValue);
+  lostItemPolicy = scoped('[data-test-lost-item-policy] [data-test-kv-value]', KeyValue);
   closeButton = scoped('button[icon=times]', ButtonInteractor);
 }